+++ src/Curiosity.SMS/EmailError.cs
@@ -1,0 +1,28 @@
+namespace Curiosity.SMS
+{
+    /// <summary>
+    /// Errors of sending SMS.
+    /// </summary>
+    public enum SmsError
+    {
+        /// <summary>
+        /// None errors.
+        /// </summary>
+        None = 0,
+
+        /// <summary>
+        /// Auth error (incorrect auth credentials, etc).
+        /// </summary>
+        Auth = 1,
+
+        /// <summary>
+        /// Network error or remote service in unavailable.
+        /// </summary>
+        Communication = 2,
+
+        /// <summary>
+        /// Unknown.
+        /// </summary>
+        Unknown = 3
+    }
+}